minor animation fix and adding layouts for the buttons
