+++ frontend/src/components/ui/LeftSide.jsx
@@ -4,8 +4,9 @@ import { ScrollTrigger } from 'gsap/ScrollTrigger';
 import { MotionPathPlugin } from 'gsap/all';  
 import bookImage from '../../assets/book_empty.png';
 import GenreTags from '../../components/GenreTags';
-import ActionButtons from '../ActionButtons';
+import ActionButton from '../ActionButton';
 import { Elastic } from 'gsap';
+import ActionButtons from '../ActionButtons';
 
 const LeftSide = ({topOneBook}) => {
   // const { title } = topOneBook.volumeInfo;
@@ -93,8 +94,11 @@ const LeftSide = ({topOneBook}) => {
   return (
     // outer color : --color-base ?
     // inner color : --color-base ?
-      <div style={{boxShadow : "-18px 20px 25px -16px rgba(0,0,0,0.58)"}} className='sidebar py-1 font-inter text-2xl opacity-0 max-w-[620px] text-black bg-[var(--color-base)] rounded-4xl h-[420px]' onMouseEnter={() => setIsHovered(true)} onMouseLeave={() => setIsHovered(false)}>
-        <ActionButtons hover={hover} WidgetRef={sideBarRef} Ypos={-50}/>
+      <div style={{boxShadow : "-18px 20px 25px -16px rgba(0,0,0,0.58)"}} className='sidebar py-1 font-inter text-2xl opacity-0 max-w-[620px] text-black bg-[var(--color-base)] rounded-4xl h-[420px] relative' onMouseEnter={() => setIsHovered(true)} onMouseLeave={() => setIsHovered(false)}>
+        <ActionButtons Ypos={-50.5} hover={hover}>
+          <ActionButton hover={hover} WidgetRef={sideBarRef} Ypos={-50}/>
+          <ActionButton hover={hover} WidgetRef={sideBarRef} Ypos={-50}/>
+        </ActionButtons>
         <div className='main-content-div flex flex-col justify-start gap-4 p-4 '>
           <div className='flex first-content gap-2 bg-[var(--color-dark)] text-[var(--color-lighter)] px-4 py-3 rounded-4xl shadow-2xl transition-all duration-200 max-h-[370px]' ref={sideBarRef}>
             <div className='pic-div max-h-[220px] justify-center flex object-cover'>
