+++ frontend/src/components/ui/RightSide.jsx
@@ -8,7 +8,7 @@ import bookImage from '../../assets/book_empty.png';
 import GenreTags from '../../components/GenreTags';
 import BookResultsGrid from '../../components/BookResultsGrid';
 import { hover, motion } from 'framer-motion';
-import ActionButtons from '../ActionButtons';
+import ActionButton from '../ActionButton';
 const RightSide = ({ topTwoBook , topThreeBook }) => {
   
   const book2VolumeInfo = topTwoBook?.volumeInfo;
@@ -110,7 +110,7 @@ const RightSide = ({ topTwoBook , topThreeBook }) => {
         // transition={{ type : "spring" , bounce : 0.4 , duration : 1}}
       >
         <div className='right-sidebar p-4 text-[100px] max-w-[620px] bg-[var(--color-base)] rounded-2xl flex gap-2 shadow-custom opacity-0' onMouseEnter={() => setIsHovered(true)} onMouseLeave={() => setIsHovered(false)}>
-          <ActionButtons hover={hovered} WidgetRef={rightSideWidget} Ypos={-54}/>
+          <ActionButton hover={hovered} WidgetRef={rightSideWidget} Ypos={-54}/>
         <div className='flex first-content gap-3 bg-[var(--color-dark)] p-4 rounded-4xl hover:shadow-custom2 hover:-translate-y-1 transition-all duration-200 text-[var(--color-lighter)]' ref={rightSideWidget}>
           <div className='pic-div-b2 max-h-[220px] justify-center flex'>
             <img src={imglink2 || bookImage} alt=""  className='min-w-[100px] rounded-2xl ring-1 object-cover'/>
